requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.11
langchain-community==0.0.25
langchain-huggingface==0.0.3
langchain-core>=0.1.30
langchain-text-splitters>=0.0.1
huggingface-hub>=0.23.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
pypdf==4.1.0
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.1
pydantic>=2.7.0
tenacity>=8.2.0
